+++ src/routes/blob_upload.rs
@@ -5,18 +5,17 @@ use axum::body::Body;
 use axum::extract::{Path, Query, State};
 use axum::response::Response;
 use axum::routing::{post, put};
-use digest::Digest;
 use hyper::StatusCode;
 
 use super::macros::endpoint_fn_7_levels;
 use crate::TrowServerState;
-use crate::registry::server::PROXY_DIR;
-use crate::registry::{ContentInfo, digest};
+use crate::registry::{ContentInfo, PROXY_DIR};
 use crate::routes::macros::route_7_levels;
 use crate::routes::response::errors::Error;
 use crate::routes::response::trow_token::TrowToken;
 use crate::routes::response::upload_info::UploadInfo;
 use crate::types::{AcceptedUpload, DigestQuery, OptionalDigestQuery, Upload};
+use crate::utils::digest::Digest;
 
 mod utils {
     use std::ops::RangeInclusive;
@@ -44,13 +43,11 @@ mod utils {
         let upload_id_bin = Uuid::parse_str(upload_id).unwrap();
 
         let size = state
-            .registry
             .storage
             .write_blob_part_stream(&upload_id_bin, data.into_data_stream(), range)
             .await?;
 
         state
-            .registry
             .storage
             .complete_blob_write(&upload_id_bin, digest.as_str())
             .await?;
@@ -180,7 +177,6 @@ async fn patch_blob_upload(
 
     let content_range = content_info.map(|ci| ci.range.0..=ci.range.1);
     let size = state
-        .registry
         .storage
         .write_blob_part_stream(&uuid, chunk.into_data_stream(), content_range)
         .await?;
@@ -340,7 +336,6 @@ mod tests {
     use uuid::Uuid;
 
     use super::*;
-    use crate::registry::Digest;
     use crate::test_utilities::{self, resp_header};
 
     // POST blob upload
@@ -472,7 +467,6 @@ mod tests {
         .await
         .unwrap();
         state
-            .registry
             .storage
             .write_blob_part_stream(&upload_uuid, Body::from("whazaaa").into_data_stream(), None)
             .await