+++ src/routes/health.rs
@@ -9,9 +9,9 @@ use crate::registry::api_types::HealthStatus;
 * GET /healthz
 */
 
-pub async fn healthz(State(state): State<Arc<TrowServerState>>) -> HealthStatus {
+pub async fn healthz(State(_state): State<Arc<TrowServerState>>) -> HealthStatus {
     HealthStatus {
         message: "".to_string(),
-        is_healthy: state.registry.is_healthy().await,
+        is_healthy: true,
     }
 }